Внесены правки по замечаниям ревьюера, а также все классы *Handler.java собраны в новый пакет handlers.
