+++ src/tracker/server/HttpTaskServer.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import com.sun.net.httpserver.HttpServer;
 import tracker.controllers.Managers;
 import tracker.controllers.TaskManager;
+import tracker.server.handlers.*;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
