+++ src/tracker/server/adapters/LocalDateTimeAdapter.java
@@ -1,0 +1,24 @@
+package tracker.server.adapters;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+import tracker.model.Task;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
+    private static final DateTimeFormatter dtf = Task.DATE_FORMATTER;
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final LocalDateTime localDateTime) throws IOException {
+        jsonWriter.value(localDateTime.format(dtf));
+    }
+
+    @Override
+    public LocalDateTime read(final JsonReader jsonReader) throws IOException {
+        return LocalDateTime.parse(jsonReader.nextString(), dtf);
+    }
+}
