+++ src/tracker/server/handlers/TasksHandler.java
@@ -1,9 +1,8 @@
-package tracker.server;
+package tracker.server.handlers;
 
 import com.google.gson.JsonParser;
 import com.google.gson.JsonSyntaxException;
 import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
 import tracker.controllers.TaskManager;
 import tracker.exceptions.ManagerSaveException;
 import tracker.exceptions.NotFoundException;
@@ -14,40 +13,13 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 
-public class TasksHandler extends BaseHttpHandler implements HttpHandler {
+public class TasksHandler extends BaseHttpHandler {
     public TasksHandler(TaskManager taskManager) {
-        super(taskManager);
+        this.taskManager = taskManager;
     }
 
     @Override
-    public void handle(HttpExchange exchange) throws IOException {
-        //System.out.println("Началась обработка /tasks запроса от клиента.");
-
-        Endpoint endpoint = getEndpoint(exchange.getRequestURI().getPath(), exchange.getRequestMethod());
-
-        switch (endpoint) {
-            case GET_TASKS: {
-                handleGetTasks(exchange);
-                break;
-            }
-            case GET_TASK: {
-                handleGetTask(exchange);
-                break;
-            }
-            case POST_TASK: {
-                handlePostTask(exchange);
-                break;
-            }
-            case DELETE_TASK: {
-                handleDeleteTask(exchange);
-                break;
-            }
-            default:
-                sendNotFound(exchange, "Такого эндпоинта не существует.");
-        }
-    }
-
-    private void handleGetTasks(HttpExchange exchange) throws IOException {
+    void handleGetTasks(HttpExchange exchange) throws IOException {
         List<Task> taskList = taskManager.getTasks();
         if (taskList.isEmpty()) {
             sendNotFound(exchange, "Список задач пуст.");
@@ -56,7 +28,8 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
         sendText(exchange, GSON.toJson(taskList));
     }
 
-    private void handleGetTask(HttpExchange exchange) throws IOException {
+    @Override
+    void handleGetTask(HttpExchange exchange) throws IOException {
         try {
             int taskID = Integer.parseInt(exchange.getRequestURI().getPath().split("/")[2]);
             Task task = taskManager.getTaskByID(taskID);
@@ -68,7 +41,8 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
         }
     }
 
-    private void handlePostTask(HttpExchange exchange) throws IOException {
+    @Override
+    void handlePostTask(HttpExchange exchange) throws IOException {
         try {
             String stringTask = new String(exchange.getRequestBody().readAllBytes(), StandardCharsets.UTF_8);
             if (!JsonParser.parseString(stringTask).isJsonObject()) {
@@ -97,7 +71,8 @@ public class TasksHandler extends BaseHttpHandler implements HttpHandler {
         }
     }
 
-    private void handleDeleteTask(HttpExchange exchange) throws IOException {
+    @Override
+    void handleDeleteTask(HttpExchange exchange) throws IOException {
         try {
             int taskID = Integer.parseInt(exchange.getRequestURI().getPath().split("/")[2]);
             taskManager.deleteTask(taskID);
